Make graphite depend on apache module
Previously, graphite managed its own apache related
resources (package/service).

Since we need to deploy it with other applications that
also depend on apache, we need to decouple the apache
resources and instead depend on the apache module.

This commit replaces the resource declarations with include
statements that use the puppetlabs apache module to configure
apache.