Update ports management
This commit updates how graphite manages
ports.conf so that it can work with newer versions
of the apache module.

In newer versions of the apache module, the ports.conf
file is managed as fragments. This resulted in duplicate
definition failures b/c graphite was also trying to manage
the file.

This commit updates the config class to use the new defined
resources types that are used to manage port entries for
apache.